+++ NEWS.md
@@ -1,8 +1,12 @@
 # markovmix 0.1.1
 
-## Enhancements:
+## New features:
 
 * Add `[.MarkovMix()` and `[<-.MarkovMix()` methods to extract and replace components in a `MarkovMix` object.
+
+## Enhancements:
+
+* Add examples of utility functions to `README.Rmd`.
 
 * Add `pillar` explicitly to imports, since it is mentioned in the documentation of `print.MarkovMix()` function.